Add sanity check for base observations (#54)
Add sanity check for base observations.  If the base observations start getting too old to be used (>= 10s) -- invoke an (optionally provided) callback to indicate this.  In theory this should never happen (TCP congestion avoidance should prevent it) but if something upstream is buffering too many RTCM messages this callback provides a hook for a user of this library to do something about it.