Upgrade to Xcode 10, pin Rust nightly [INFRA-202] (#402)
Attempting to resolve build errors here: https://travis-ci.org/github/swift-nav/gnss-converters/jobs/692078797

Something is wonky with xcode8 nodes on Travis... xcode8 is from 2016, so we can just bump to xcode10 (release in 2018) to fix this issue.  Xcode 10 on Travis is macOS version 10.12.

Also pin the nightly version in case this was an issue with Rust nightly dropping support for older macOS versions.